🎨 New design for list of groups
